backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.1
django-filter>=24.2
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
whitenoise>=6.6
django-csp>=3.7
drf-spectacular>=0.27
pytest>=8.0
pytest-django>=4.8
django-cors-headers>=4.9.0